packages/orq-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orq-mcp"
version = "0.1.0"
description = "MCP server for Orq AI platform integration"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Orq AI", email = "support@orq.ai"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=2.12.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]

[project.urls]
Homepage = "https://github.com/orq-ai/orqkit"
Repository = "https://github.com/orq-ai/orqkit"

[tool.hatch.build.targets.wheel]
packages = ["src/orq_mcp"]
